+++ src/main/java/com/evaluasi/EvaluasiHUMBackEnd/entity/Cpt.java
@@ -20,7 +20,7 @@ public class Cpt {
     private int panolcustomer;
     private int coverage;
     private double coveragepersen;
-    private int penetration;
+    private double penetration;
     private double throughput;
     private double hitrate;
